Add Gemfile.lock to gitignore
Yehuda put it more clearly than I can, but having Gemfile.lock checked
into a library as opposed to an application enforces a level of
strictness that isn't necessary. With the version specifications for the
gem's dependencies present in the .gemspec, theoretically bundling any
latest patch version of those gems should be fine.

Signed-off-by: David Celis <me@davidcel.is>