Refactor incrementer code
Now that custom incrementers are no longer allowed, the incrementer code
can be refactored to be simpler. No longer do we need to save a Proc to
handle incrementer logic. Replaced with a simple method named step that
provides the size value for a particular n. n = 0 is min_size.