Rename chart attribute to chart_file
The @chart attribute in Job was holding the value of the file to write
out the chart data to, which is a poor name. Changed this to chart_file
to match the conventions of json_file and csv_file, and to be overall a
better description. Also remove reader for this attribute.