Add Chart class
Create a Chart class to hold chart-related knowledge. At this point it
is just a basic stub, and just the basic chart writing code from Job is
copied over into a generate method. Job creates a Chart and passes a
reference to Report and the sizes object. Further refactoring needed.